refactor detail screen
