+++ app/src/main/java/com/example/zeldacompendium/presentation/ui/detail/sections/DetailViewModel.kt
@@ -8,7 +8,7 @@ import dagger.hilt.android.lifecycle.HiltViewModel
 import javax.inject.Inject
 
 @HiltViewModel
-class ItemDetailViewModel @Inject constructor(
+class DetailViewModel @Inject constructor(
    private val repository: CompendiumRepository
 ) : ViewModel() {
