+++ app/src/main/java/com/example/zeldacompendium/presentation/ui/detail/sections/categories/equipment/EquipmentsInfo.kt
@@ -1,16 +1,16 @@
 @file:OptIn(ExperimentalFoundationApi::class)
-package com.example.zeldacompendium.presentation.ui.detail.sections.categories.treasures
+package com.example.zeldacompendium.presentation.ui.detail.sections.categories.equipment
 
 import androidx.compose.foundation.ExperimentalFoundationApi
 import androidx.compose.foundation.pager.rememberPagerState
 import androidx.compose.runtime.Composable
 import com.example.zeldacompendium.data.remote.responses.ItemDetailModel
 import com.example.zeldacompendium.presentation.ui.components.horizontalpager.sections.pager.CustomHorizontalPager
-import com.example.zeldacompendium.presentation.ui.components.horizontalpager.sections.page.type.DropsPage
 import com.example.zeldacompendium.presentation.ui.components.horizontalpager.sections.page.type.LocationPage
+import com.example.zeldacompendium.presentation.ui.components.horizontalpager.sections.page.type.PropertiesPage
 
 @Composable
-fun TreasureItemDetailContainer(
+fun EquipmentsInfo(
    itemInfo: ItemDetailModel,
    gameId: Int
 ) {
@@ -18,9 +18,9 @@ fun TreasureItemDetailContainer(
    CustomHorizontalPager(
       pagerState = pagerState,
    ) { page ->
-      when (page) {
-         0 -> DropsPage(itemInfo)
-         1 -> LocationPage(itemInfo, gameId)
-      }
+         when (page) {
+            0 -> PropertiesPage(itemInfo, gameId)
+            1 -> LocationPage(itemInfo, gameId)
+         }
    }
 }
